Ported a number of posts and added vimeo plugin.
I added a vimeo plugin so that I could post vimeo videos in my blog
posts. It took some googling to figure out how to make the vimeo snippet
valid in terms of the Markdown parser that jekyll uses. It was pretty
simple, I just had to assign the attributes that didn't have values,
values, and put some content inside the empty iframe tags because it
seems to cause failure to render the rest of the post when the iframe is
empty. The vimeo plugin I took from the web and hacked a bit also
provides consistency of the size and format in all of my blog posts.