Added little post creation bash script.
I did this to ease the initial creation of the posts. I based it on a
couple of other scripts that I saw people had developed on the web.
Basically, it lets you give an optional title and optional date and
creates the file with the proper YAML header and opens up vim to edit
the post.