Added a medium_photo tag, used it in a post.
I added a liquid tag called medium_photo which is used to place my
medium sized photos inside blog posts. It simply takes as the first
argument the name of the file, ex: dsc_0432.jpg and for the second
argument which is optional it takes an align attribute value, left,
right, top, bottom.